Move function to spectrum (mismatch.py)
