+++ docs/sphinx/source/whatsnew/v0.11.0.rst
@@ -17,7 +17,7 @@ Enhancements
 ~~~~~~~~~~~~
 * Add module :py:mod:`pvlib.par` to facilitate the design of Agrivoltaic
   systems.
-  - Adds function :py:func:`pvlib.par.spitters_relationship` to calculate the
+  - Adds function :py:func:`pvlib.spectrum.spitters_relationship` to calculate the
     diffuse fraction of Photosynthetically Active Radiation (PAR) from the
     global diffuse fraction and the solar zenith.
     (:issue:`2047`, :pull:`2048`)
