+++ pvlib/__init__.py
@@ -14,7 +14,6 @@ from pvlib import (  # noqa: F401
     ivtools,
     location,
     modelchain,
-    par,
     pvarray,
     pvsystem,
     scaling,
